# api/swagger/marketplace.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Marketplace API",
    "version": "1.0"
  },
  "host": "marketplace.dev.videocoin.network",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/account": {
      "get": {
        "summary": "Get account info",
        "operationId": "GetAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AccountResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "tags": [
          "Proifle"
        ]
      },
      "put": {
        "summary": "Update account info",
        "operationId": "UpdateAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AccountResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateAccountRequest"
            }
          }
        ],
        "tags": [
          "Proifle"
        ]
      }
    },
...
    "/api/v1/assets": {
      "get": {
        "summary": "Get list of assets",
        "operationId": "GetAssets",
        "security": [],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AssetsResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "tags": [
          "Assets"
        ]
      },
      "post": {
        "summary": "Create asset",
        "operationId": "CreateAsset",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AssetResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateAssetRequest"
            }
          }
        ],
        "tags": [
          "Assets"
        ]
      }
    },
...
    "/api/v1/assets/{id}": {
      "get": {
        "summary": "Get detail asset info",
        "operationId": "GetAsset",
        "security": [],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AssetResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "format": "integer"
          }
        ],
        "tags": [
          "Assets"
        ]
      }
    },
    "/api/v1/asset/{contract_address}/{token_id}": {
      "get": {
        "summary": "Get detail asset info by contract address and token id",
        "operationId": "GetAssetByContractAddressAndTokenID",
        "security": [],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AssetResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "parameters": [
          {
            "name": "contract_address",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Assets"
        ]
      }
    },
...
    "/api/v1/creators/{creator_id}/assets": {
      "get": {
        "summary": "Get list of assets by creator",
        "operationId": "GetAssetsByCreator",
        "security": [],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AssetsResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "parameters": [
          {
            "name": "creator_id",
            "in": "path",
            "required": true,
            "type": "number",
            "format": "integer"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "integer"
          }
        ],
        "tags": [
          "Assets"
        ]
      }
    },
...
  },
  "definitions": {
...
    "CreateAssetRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "asset_id": {
          "type": "number",
          "format": "integer"
        },
        "desc": {
          "type": "string"
        },
        "yt_video_link": {
          "type": "string"
        },
        "on_sale": {
          "type": "boolean"
        },
        "royalty": {
          "type": "number",
          "format": "integer"
        },
        "instant_sale_price": {
          "type": "number",
          "format": "float"
        }
      }
    },
...
    "UpdateAccountRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "custom_url": {
          "type": "string"
        },
        "yt_username": {
          "type": "string"
        },
        "image_data": {
          "type": "string"
        },
        "cover_data": {
          "type": "string"
        }
      }
    },
...
    "UserResponse": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "cover_url": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "custom_url": {
          "type": "string"
        },
        "yt_username": {
          "type": "string"
        }
      }
    },
    "AccountResponse": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "profile_img_url": {
          "type": "string"
        },
        "is_verified": {
          "type": "boolean"
        },
        "user": {
          "$ref": "#/definitions/UserResponse"
        }
      }
    },
...
    "AssetContractResponse": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "asset_contract_type": {
          "type": "string"
        },
        "schema_name": {
          "type": "string"
        },
        "symbol": {
          "type": "string"
        },
        "buyer_fee_basis_points": {
          "type": "number"
        },
        "seller_fee_basis_points": {
          "type": "number"
        },
        "opensea_buyer_fee_basis_points": {
          "type": "number"
        },
        "opensea_seller_fee_basis_points": {
          "type": "number"
        },
        "dev_buyer_fee_basis_points": {
          "type": "number"
        },
        "dev_seller_fee_basis_points": {
          "type": "number"
        }
      }
    },
    "AssetResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "format": "integer"
        },
        "token_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "content_type": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "thumbnail_url": {
          "type": "string"
        },
        "preview_url": {
          "type": "string"
        },
        "encrypted_url": {
          "type": "string"
        },
        "qr_url": {
          "type": "string"
        },
        "ipfs_url": {
          "type": "string"
        },
        "ipfs_thumbnail_url": {
          "type": "string"
        },
        "ipfs_encrypted_url": {
          "type": "string"
        },
        "yt_video_id": {
          "type": "string"
        },
        "drm_key": {
          "type": "string"
        },
        "on_sale": {
          "type": "boolean"
        },
        "sold": {
          "type": "boolean"
        },
        "instant_sale_price": {
          "type": "number",
          "format": "float"
        },
        "owner": {
          "$ref": "#/definitions/AccountResponse"
        },
        "asset_contract": {
          "$ref": "#/definitions/AssetContractResponse"
        }
      }
    },
    "AssetsResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AssetResponse"
          }
        },
        "total_count": {
          "type": "number",
          "format": "integer"
        },
        "count": {
          "type": "number",
          "format": "integer"
        },
        "prev": {
          "type": "boolean",
          "format": "boolean"
        },
        "next": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}